requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==3.12.3
python-dotenv==1.0.1
django-storages==1.14.2
Pillow==10.2.0
django-cleanup==8.1.0
mongoengine==0.27.0 